pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medicine-verifier-api"
version = "1.0.0"
description = "Universal Medicine Verification API"
authors = [{name = "Medicine Verifier Team"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "opencv-python>=4.8.1",
    "pytesseract>=0.3.10",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "numpy<2.0",
    "pillow>=10.1.0",
    "aiohttp>=3.9.0",
    "fuzzywuzzy>=0.18.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
